finish migrating dynamo utility functions, fix some type annotations
